% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_add_regularization.R
\name{dpca_add_regularization}
\alias{dpca_add_regularization}
\title{A function to add regularization in the Demixed PCA
Utilized in `dpca_fit()` and `dpca_fit_transform()`.}
\usage{
dpca_add_regularization(Y, mYs, lam)
}
\arguments{
\item{Y}{An array, dim=c(n_neurons, n_features_1, n_features_2, ...), where n_samples in the number of neurons and n_features_j is the number of the j-features (where the axis correspond to different parameters).}

\item{mYs}{List with values in the shape of flattened Y. Marginalized data, should be the result of dpca_marginalize().}

\item{lam}{The coefficient for the Ridge regularization. A non-negative number. lam equal to the product of regularizer and l2 norm of the data.}
}
\value{
return the regularized flattened matrix, marginalizations, and the inversed matrix.
}
\description{
A function to add regularization in the Demixed PCA
Utilized in `dpca_fit()` and `dpca_fit_transform()`.
}
