% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_data_sim.R
\name{dpca_data_sim}
\alias{dpca_data_sim}
\title{Generate the simulation data}
\usage{
dpca_data_sim(
  N = 50,
  n_samples = 10,
  noise = 0.2,
  pars_lst,
  save = TRUE,
  file_name = "simR.rds"
)
}
\arguments{
\item{N}{number of neuron samples. Default to be 50.}

\item{n_samples}{number of trail samples. Default to be 10.}

\item{noise}{noise in the trail experiments. Default to be 0.2.}

\item{pars_lst}{A list of number containing the number of each task parameter features}

\item{save}{A bool parameter, whether to save the simulated data as `.rds` file or not.}

\item{file_name}{The name of file to be save as. Default to be 'simR.rds'}
}
\value{
A simulated neural population trail data
}
\description{
Generate the simulation data
}
