% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_decomposition.R
\name{dpca_decomposition}
\alias{dpca_decomposition}
\title{Doing the decomposition of matrix with different methods}
\usage{
dpca_decomposition(flatX, mXs, pinvX, n_components = 10, method = "rsvd")
}
\arguments{
\item{flatX}{A 2D matrix flattened from a high dimensional array. Dim = c(num_neuron, num_features).}

\item{mXs}{List with values in the shape of flatX. Marginalized data, should be the result of dpca_marginalize()}

\item{pinvX}{Inverse of the flatX matrix}

\item{n_components}{The number of components required. Default=10.}

\item{method}{The method chose for decomposition. 'rsvd': Randomized SVD, 'svd': SVD, 'qr': QR decomposition. Default='rsvd'.}
}
\value{
P: A list mapping strings to matrix. Element dim = c(num_neuron, n_components). Holds encoding matrices for each term in variance decompostions ( can be used in inverse_transform to map from low-dimensional representation back to original data space).

F: A list mapping strings to matrix. Element dim = c(num_neuron, n_components). Holds decoding matrices for each term in variance decompostions (used to transform data to low-dimensional space).
}
\description{
Doing the decomposition of matrix with different methods
}
