% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_fit.R
\name{dpca_fit}
\alias{dpca_fit}
\title{A function to fit the high dimensional data (neural population data) via Demixed PCA.}
\usage{
dpca_fit(X, n_components = 10, regularizer = 0, method = "rsvd", label_lst)
}
\arguments{
\item{X}{An array, dim=c(n_neurons, n_features_1, n_features_2, ...), where n_samples in the number of neurons and n_features_j is the number of the j-features (where the axis correspond to different parameters).}

\item{n_components}{The number of components required. Defult=10.}

\item{regularizer}{The coefficient for the Ridge regularization. A non-negative number. Defult to be 0, no regularization.}

\item{method}{The method chose for decomposition. 'rsvd': Randomized SVD, 'svd': SVD, 'qr': QR decomposition. Defult='rsvd'.}

\item{label_lst}{A list of strings presenting the task parameters. The order of element in the string should consistent with the dimension of the array data. 'ts': time and stimulus, 'dts': decision, time and stimulus, etc.}
}
\value{
P: A list mapping strings to matrix. Element dim = c(num_neuron, n_components). Holds encoding matrices for each term in variance decompostions ( can be used in inverse_transform to map from low-dimensional representation back to original data space).

F: A list mapping strings to matrix. Element dim = c(num_neuron, n_components). Holds decoding matrices for each term in variance decompostions (used to transform data to low-dimensional space).
}
\description{
A function to fit the high dimensional data (neural population data) via Demixed PCA.
}
