% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_flat2d.R
\name{dpca_flat2d}
\alias{dpca_flat2d}
\title{Flatten array function}
\usage{
dpca_flat2d(A)
}
\arguments{
\item{A}{The high-dimension array.}
}
\value{
A matrix flatten from the array
}
\description{
This function allows you to flatten a high-dimension array into a 2-dimension matrix
}
\examples{
flat2d(array(1, dim=c(3, 3, 3)))
}
