% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_marginalize.R
\name{dpca_marginalize}
\alias{dpca_marginalize}
\title{Marginalize the data matrix}
\usage{
dpca_marginalize(X, label_lst)
}
\arguments{
\item{X}{An array, dim=c(n_neurons, n_features_1, n_features_2, ...), where n_samples in the number of neurons and n_features_j is the number of the j-features (where the axis correspond to different parameters).}

\item{label_lst}{A string presenting the task parameters. The order of element in the string should consistent with the dimension of the array data. 'ts': time and stimulus, 'dts': decision, time and stimulus, etc.}
}
\value{
A list mapping strings to matrices, with values corresponding to the marginalized data and mixed residual data (and the key refers to the marginalization label, mixed residual data is called by 'mix')
}
\description{
Marginalize the data matrix
}
