% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_plot_component.R
\name{dpca_plot_component}
\alias{dpca_plot_component}
\title{Plot the `k`th components. Utilized in the `dpca_fit_transform()` function. Can handle neural population data with 3 or 4 task parameters.}
\usage{
dpca_plot_component(
  X,
  re_transform,
  k_component = 1,
  plot_margin,
  ylim = c(-3, 3)
)
}
\arguments{
\item{X}{An array, dim=c(n_neurons, n_features_1, n_features_2, ...), where n_samples in the number of neurons and n_features_j is the number of the j-features (where the axis correspond to different parameters).}

\item{re_transform}{A list mapping strings to matrix. Each element is a demixed pca transformed (decoding) 2D matrix.}

\item{k_component}{The `k`th component to plot. Default to be 1.}

\item{plot_margin}{A string presenting the margin (task parameter) to plot the components.}

\item{ylim}{The y axis limitation. Same as `plot()` function in R. Default to be c(-3,3).}
}
\value{
A plot with the component on margin
}
\description{
Plot the `k`th components. Utilized in the `dpca_fit_transform()` function. Can handle neural population data with 3 or 4 task parameters.
}
