% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_reconstruct.R
\name{dpca_reconstruct}
\alias{dpca_reconstruct}
\title{Reconstruct the data: Transform data first into reduced space before projecting it back into data space.}
\usage{
dpca_reconstruct(
  X,
  n_components = 5,
  margin_on,
  label,
  regularizer = 0,
  method = "rsvd"
)
}
\arguments{
\item{X}{An array, dim=c(n_neurons, n_features_1, n_features_2, ...), where n_samples in the number of neurons and n_features_j is the number of the j-features (where the axis correspond to different parameters).}

\item{n_components}{The number of components required. Defult=5.}

\item{margin_on}{A string presenting the margin (task parameter) to be inversed transformed on.}

\item{label}{A string presenting the task parameters. The order of element in the string should consistent with the dimension of the array data. 'ts': time and stimulus, 'dts': decision, time and stimulus, etc.}

\item{regularizer}{The coefficient for the Ridge regularization. A non-negative number. Deault to be 0, no regularization.}

\item{method}{The method chose for decomposition. 'rsvd': Randomized SVD, 'svd': SVD, 'qr': QR decomposition. Default='rsvd'.}
}
\value{
An array with the same shape as the original data.
}
\description{
Reconstruct the data: Transform data first into reduced space before projecting it back into data space.
}
