% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_transform.R
\name{dpca_transform}
\alias{dpca_transform}
\title{A function to transform the high dimensional data (neural population data) via Demixed PCA.}
\usage{
dpca_transform(X, marginalizations, P, F)
}
\arguments{
\item{X}{An array, dim=c(n_neurons, n_features_1, n_features_2, ...), where n_samples in the number of neurons and n_features_j is the number of the j-features (where the axis correspond to different parameters).}

\item{marginalizations}{String or None. Marginalization subspace upon which to project, if None return dict with projections on all marginalizations.}

\item{P}{A list mapping strings to matrix. Element dim = c(num_neuron, n_components). Holds encoding matrices for each term in variance decompositions.}

\item{F}{A list mapping strings to matrix. Element dim = c(num_neuron, n_components). Holds decoding matrices for each term in variance decompositions.}
}
\value{
A list mapping strings to matrix. Each element is a demixed pca transformed (decoding) 2D matrix.
}
\description{
A function to transform the high dimensional data (neural population data) via Demixed PCA.
}
