% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca_zero_mean.R
\name{dpca_zero_mean}
\alias{dpca_zero_mean}
\title{Subtracts the mean from each observable}
\usage{
dpca_zero_mean(A)
}
\arguments{
\item{A}{a High dimensional array}
}
\value{
a same dimension array that already subtracted the mean
}
\description{
Subtracts the mean from each observable
}
